+++ MUS4_FW/libraries/mus4_core/src/mus4_core.h
@@ -10,6 +10,7 @@
 
 #include "BuildInfo.h"
 #include "FirmwareConfig.h"
+#include "MutePreference.h"
 #include "SharedTypes.h"
 #include "RuntimeState.h"
 #include "WifiConsoleTypes.h"